B1.3: the launch composer grows sideways at its line-group seam
S3 of research/lens-keying-successors-2026-09-02.md. `_command`,
`_identity_line`, `_lane_lines`, `_reading_lines`, `_files_findings`,
`_craft_lines`, `_friction_lines`, `_return_lines`, `ARTIFACT_LINE_FORMS`
and `FINDINGS_LINE` move byte-identical to
`scripts/tickets_dispatch_launch_lines.py`, which carries the same
`if __package__:` flat-import fallback every scripts module does.
`tickets_dispatch_launch.py` keeps `launch_prompt`, the host binding
functions, `launch_spec` and `precheck`, and its docstring now points at
the new owner for what the child-facing surface may carry.

The one moved symbol imported elsewhere, `ARTIFACT_LINE_FORMS` in
`tests/test_ticket_callables.py`, imports from the new owner; nothing is
re-exported. `ReturnLineConditionalTest`, `IdentityLineTest` and
`LensKeyPromptTest` move to `tests/test_dispatch_launch_lines.py`;
`LaunchResolutionTest`, `DispatchLaunchTest` and `LandTest` stay
byte-identical in place, S4's `LandTest` case included.
