+++ ARCHITECTURE.md
@@ -43,7 +43,9 @@ family's five owner modules.
   `tickets_emission.py` emission grading; `tickets_issue_render.py`
   issuance markdown; `tickets_mint.py` and `tickets_frame.py` the callable
   and frame minting commands;
-  `tickets_dispatch_launch.py` resolves the host launch binding. `cutcheck.py`
+  `tickets_dispatch_launch.py` resolves the host launch binding and orders the
+  child prompt, whose line groups are `tickets_dispatch_launch_lines.py`'s.
+  `cutcheck.py`
   owns structural graph validation.
   Cutcheck imports those owners directly, never the tickets facade;
   admission and cutcheck never import each other.
